DecryptCode/requirements.txt
boto3
python-gnupg